feat : script to clean the sales details in silver layer

+++ scripts/silver/ddl_silver.sql
@@ -59,9 +59,9 @@ CREATE TABLE silver.crm_sales_details (
     sls_order_dt DATE,
     sls_ship_dt DATE,
     sls_due_dt DATE,
-    sls_sales DECIMAL(18,2),
+    sls_sales INT,
     sls_quantity INT,
-    sls_price DECIMAL(18,2),
+    sls_price INT,
     dwh_create_date DATETIME2 DEFAULT GETDATE()
 );
 GO
